# backend/app/swagger_ui/swagger.yml
openapi: 3.0.0
info:
  title: Sample API
  version: 1.0.0
...
servers:
  - url: http://127.0.0.1:5000/
paths:
...
  /api/configuration/{user_id}/{equipment_id}:
    get:
      summary: Получить Configuration для одного пользователя
      tags:
        - Configuration
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID User
        - name: equipment_id
          in: path
          required: true
          schema:
            type: integer
          description: ID equipment
      responses:
        '200':
          description: Configuration для одного пользователя
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: [ ]
    delete:
      summary: Удалить Configuration
      tags:
        - Configuration
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID User
        - name: equipment_id
          in: path
          required: true
          schema:
            type: integer
          description: ID equipment
      responses:
        '200':
          description: Результат
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: [ ]
    post:
      summary: Добавить Configuration для одного пользователя
      tags:
        - Configuration
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID User
        - name: equipment_id
          in: path
          required: true
          schema:
            type: integer
          description: ID equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: string
                  description: json config
              required:
                - config
      responses:
        '200':
          description: Configuration для одного пользователя
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: [ ]
    put:
      summary: Изменить Configuration для одного пользователя
      tags:
        - Configuration
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID User
        - name: equipment_id
          in: path
          required: true
          schema:
            type: integer
          description: ID equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: string
                  description: json config
              required:
                - config
      responses:
        '200':
          description: Configuration для одного пользователя
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: [ ]
  /api/configuration/{user_id}/{equipment_id}/apply:
    get:
      summary: Применить Configuration для одного пользователя
      tags:
        - Configuration
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID User
        - name: equipment_id
          in: path
          required: true
          schema:
            type: integer
          description: ID equipment
        - name: last_update
          in: query
          schema:
            type: string
            format: date-time
          description: Начальная дата в формате YYYY-MM-DD
          required: true
      responses:
        '200':
          description: Список значений Sensor_records
          content:
            application/json:
              schema:
                type: integer
      security:
        - BearerAuth: [ ]
  /api/configuration/functions:
    get:
      summary: Получить функции
      tags:
        - Configuration
      responses:
        '200':
          description: Список функции
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: [ ]
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT